% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{BiocFERContainer-class}
\alias{BiocFERContainer-class}
\alias{.BiocFERContainer}
\title{The Bioconductor Foreach Redis container}
\description{
The Bioconductor Foreach Redis container. See \code{?BiocFERServerContainer} and
\code{?BiocFERWorkerContainer}
}

