% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-BiocFERContainer.R
\name{BiocFERContainer-commom-parameters}
\alias{BiocFERContainer-commom-parameters}
\title{Common BiocFERContainer parameter}
\arguments{
\item{image}{Character, the container image}

\item{name}{Character, the optional name of the container}

\item{environment}{List, the environment variables in the container}
}
\description{
Common BiocFERContainer parameter
}
