% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-BiocFERContainer.R
\name{BiocFERServerContainer}
\alias{BiocFERServerContainer}
\title{Get the Bioconductor Foreach Redis server container}
\usage{
BiocFERServerContainer(
  image = "dockerparallel/parallel-redis-server",
  name = "redisRServerContainer",
  environment = list()
)
}
\arguments{
\item{image}{Character, the container image}

\item{name}{Character, the optional name of the container}

\item{environment}{List, the environment variables in the container}
}
\value{
a \code{BiocFERContainer} object
}
\description{
Get the Bioconductor Foreach Redis server container.
}
\examples{
BiocFERServerContainer()
}
