% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ClusterMethodGetter.R
\name{$,ClusterMethodGetter-method}
\alias{$,ClusterMethodGetter-method}
\title{Get the exported object by the name}
\usage{
\S4method{$}{ClusterMethodGetter}(x, name)
}
\arguments{
\item{x}{\code{ClusterMethodGetter} object}

\item{name}{Character name}
}
\value{
the exported object
}
\description{
Get the exported object by the name
}
