% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-DockerCluster.R
\name{$,DockerCluster-method}
\alias{$,DockerCluster-method}
\title{Get the exported object}
\usage{
\S4method{$}{DockerCluster}(x, name)
}
\arguments{
\item{x}{The \code{DockerCluster} object}

\item{name}{Character, the name of the exported object}
}
\value{
The object in the cluster
}
\description{
Get the exported object
}
