% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-DockerCluster.R
\name{$<-,DockerCluster-method}
\alias{$<-,DockerCluster-method}
\title{Set the value of the exported object}
\usage{
\S4method{$}{DockerCluster}(x, name) <- value
}
\arguments{
\item{x}{The \code{DockerCluster} object}

\item{name}{Character, the name of the exported object}

\item{value}{The value of the exported object}
}
\value{
The \code{DockerCluster} object
}
\description{
Set the value of the exported object
}
