% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics-CloudProvider.R,
%   R/method-default.R
\name{cleanupDockerCluster}
\alias{cleanupDockerCluster}
\alias{cleanupDockerCluster,ANY-method}
\title{Cleanup the resources after the cluster has been stopped}
\usage{
cleanupDockerCluster(provider, cluster, deep, verbose)

\S4method{cleanupDockerCluster}{ANY}(provider, cluster, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{deep}{Logical(1), wheter all the associated resources should be removed}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
Cleanup the resources after the cluster has been stopped. After this
function is called, all the non-free resources should be stopped.
The cloud provider can still preserve some resources if they
are free. This generic might be called multiple times.
The default method does nothing.
}
\section{Functions}{
\itemize{
\item \code{cleanupDockerCluster,ANY-method}: The default method, do nothing.
}}

