% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{CloudHardware-class}
\alias{CloudHardware-class}
\alias{.CloudHardware}
\title{The instance hardware for running the docker}
\description{
The instance hardware for running the docker
}
\section{Slots}{

\describe{
\item{\code{cpu}}{The CPU limitation for the docker. 1024 CPU unit
corresponds to 1 core.}

\item{\code{memory}}{The memory limitation for the docker, the unit
is MB}

\item{\code{id}}{The id of the hardware, the meaning of \code{id} depends on
the cloud provider.}
}}

