% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{CloudRuntime-class}
\alias{CloudRuntime-class}
\alias{.CloudRuntime}
\title{The cloud runtime}
\description{
The cloud runtime. It is a class purely for storing the runtime information
for the cloud.
The values in \code{CloudRuntime} in a cluster can be accessed by the getter function
which starts with the prefix \code{.get}(e.g. \code{.getServerHandle(cluster)}).
}
\section{Fields}{

\describe{
\item{\code{serverHandle}}{Any data type, the server handle that can be recognized by the
cloud provider.}

\item{\code{workerHandles}}{A list object for internal use only.
Please call \code{.getWorkerHandles(cluster)} to access the worker handles}

\item{\code{workerPerHandle}}{An internal counter. Please call \code{.getWorkerHandles(cluster)}
to access the worker handles}

\item{\code{serverPublicIp}}{Character or NULL, the server public IP}

\item{\code{serverPrivateIp}}{Character or NULL, the server private IP}
}}


