% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{ClusterMethodGetter-class}
\alias{ClusterMethodGetter-class}
\alias{.ClusterMethodGetter}
\title{An utility class}
\description{
An utility class for exporting the APIs from the cloud provider and
container.
}
