% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-export.R
\name{clusterPreset}
\alias{clusterPreset}
\title{Set the default cloud provider and container}
\usage{
clusterPreset(
  cloudProvider = c("ECSFargateProvider"),
  container = c("BiocFEDRContainer")
)
}
\arguments{
\item{cloudProvider}{The default cloud provider name, can be abbreviated}

\item{container}{The default container name, can be abbreviated}
}
\description{
Set the default cloud provider and container
}
\examples{
clusterPreset(cloudProvider = "ECSFargateProvider", container = "BiocFEDRContainer")
cluster <- makeDockerCluster()
cluster
}
