% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{commonParams}
\alias{commonParams}
\title{commom params}
\arguments{
\item{verbose}{Integer. The verbose level, default 1}

\item{provider}{S4 CloudProvider object. The service provider}

\item{cluster}{S4 DockerCluster object}

\item{container}{S4 Container Object}

\item{hardware}{S4 CloudHardware Object}
}
\description{
commom params
}
