% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{configServerContainerEnv}
\alias{configServerContainerEnv}
\title{Configurate the server container environment}
\usage{
configServerContainerEnv(container, cluster, verbose)
}
\arguments{
\item{container}{Reference Container Object. The server container.}

\item{cluster}{S4 DockerCluster object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A \code{container} object
}
\description{
Configurate the server container environment. Developers can use this function
to set the server password, port number and etc. via the container environment variable.
}
