% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BiocFERContainer.R
\name{configWorkerContainerEnv,BiocFERContainer-method}
\alias{configWorkerContainerEnv,BiocFERContainer-method}
\title{Configure the worker container}
\usage{
\S4method{configWorkerContainerEnv}{BiocFERContainer}(container, cluster, workerNumber, verbose = FALSE)
}
\arguments{
\item{container}{Reference Container Object. The worker container.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{workerNumber}{Integer. The number of workers in a container.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A \code{BiocFERContainer} object
}
\description{
Configure the worker container
}
