% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{configWorkerContainerEnv}
\alias{configWorkerContainerEnv}
\title{Configurate the worker container environment}
\usage{
configWorkerContainerEnv(container, cluster, workerNumber, verbose)
}
\arguments{
\item{container}{Reference Container Object. The worker container.}

\item{cluster}{S4 DockerCluster object.}

\item{workerNumber}{Integer. The number of workers in a container.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A \code{container} object
}
\description{
Configurate the worker container environment. Developers can use this function
to set the server Ip, password and etc. via the container environment variable.
}
