% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DockerParallel-docker_cluster.R
\name{docker_cluster}
\alias{docker_cluster}
\title{Create a cluster with remote docker workers}
\usage{
docker_cluster(config, workers, ...)
}
\arguments{
\item{config}{An S4 object, the configuration of the worker nodes}

\item{workers}{The number of workers in the cluster, additional meaning may
exist for a specific implementation}

\item{...}{Additional arguments that will passed to the specific implementation}
}
\value{
The return value is implementation dependent
}
\description{
Create a cluster with remote docker workers. This is an S4 generic.
The method will be dispatched based on the \code{config} variable
}
