% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{DockerCluster-class}
\alias{DockerCluster-class}
\alias{.DockerCluster}
\title{The docker cluster class}
\description{
The docker cluster class.
The values in the cluster can be accessed by the getter or setter function
which starts with the prefix \code{.get} or \code{.set}(e.g. \code{.getJobQueueName(cluster)}).
}
\section{Slots}{

\describe{
\item{\code{cloudProvider}}{CloudProvider}

\item{\code{cloudConfig}}{CloudConfig}

\item{\code{serverContainer}}{The container definition for the server.}

\item{\code{workerContainer}}{The container definition for the worker}

\item{\code{cloudRuntime}}{CloudRuntime}

\item{\code{settings}}{Environment, the cluster settings}
}}

