% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-DockerCluster.R
\name{DockerCluster-common-parameters}
\alias{DockerCluster-common-parameters}
\title{Common \code{DockerCluster} parameter}
\arguments{
\item{x}{The \code{DockerCluster} object}

\item{name}{Character, the name of the exported object}

\item{object}{The \code{DockerCluster} object}
}
\value{
No return value
}
\description{
Common \code{DockerCluster} parameter
}
