% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ECSFargateProvider.R
\name{dockerClusterExists,ECSFargateProvider-method}
\alias{dockerClusterExists,ECSFargateProvider-method}
\title{Whether the cluster is running on the cloud?}
\usage{
\S4method{dockerClusterExists}{ECSFargateProvider}(provider, cluster, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A logical value
}
\description{
Whether the cluster is running on the cloud?
This function will check the instances in the ECS cluster defined by
\code{provider$clusterName} and find if there is any instance that has the same job queue
name as \code{.getJobQueueName(cluster)}
}
