% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{DockerContainer-class}
\alias{DockerContainer-class}
\alias{.DockerContainer}
\title{The root class of the container}
\description{
The root class of the container
}
\section{Fields}{

\describe{
\item{\code{name}}{Character, the optional name of a container}

\item{\code{maxWorkerNum}}{Integer, the maximum worker number in a container}

\item{\code{environment}}{List, the environment variables in the container}

\item{\code{image}}{Character, the container image}
}}


