% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DockerHardware.R
\name{DockerHardware}
\alias{DockerHardware}
\title{Make a DockerHardware object}
\usage{
DockerHardware(cpu = 256, memory = 512, id = character(0))
}
\arguments{
\item{cpu}{Numeric(1), the CPU limitation for the docker. 1024 CPU unit
corresponds to 1 core.}

\item{memory}{Numeric(1), the memory limitation for the docker, the unit
is MB.}

\item{id}{character(1) or character(0), the id of the hardware, the meaning of
\code{id} depends on the cloud provider.}
}
\value{
A DockerHardware object
}
\description{
Make a DockerHardware object
}
\examples{
DockerHardware()
}
