% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics-CloudProvider.R
\name{runDockerServer}
\alias{runDockerServer}
\alias{stopDockerServer}
\title{Run or stop the server container}
\usage{
runDockerServer(provider, cluster, container, hardware, verbose)

stopDockerServer(provider, cluster, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{container}{S4 \code{DockerContainer} Object. The server container.}

\item{hardware}{S4 \code{DockerHardware} Object. The server hardware.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value, if error occurs, the function can throw an error.
}
\description{
Run or stop the server. These functions will not be called if the server is
not managed by the provider. There is no default method for these generics.
}
