% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyContainer.R
\name{DummyWorkerContainer}
\alias{DummyWorkerContainer}
\alias{DummyServerContainer}
\title{A dummy container}
\usage{
DummyWorkerContainer(
  image = "workerImage",
  backend = "testBackend",
  maxWorkerNum = 123L
)

DummyServerContainer(image = "serverImage", backend = "testBackend")
}
\arguments{
\item{image}{The image for the container}

\item{backend}{The parallel backend for the container}

\item{maxWorkerNum}{The maximum worker number}
}
\description{
A dummy container. It is for purely testing purpose.
}
\examples{
DummyWorkerContainer()
}
