% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{ECSFargateProvider-class}
\alias{ECSFargateProvider-class}
\alias{.ECSFargateProvider}
\title{The ECS fargate provider}
\description{
The ECS fargate provider, see \code{?ECSFargateProvider}
}
\section{Slots}{

\describe{
\item{\code{clusterName}}{Character, the name of the ECS cluster}

\item{\code{initialized}}{Logical, whether the provider has been initialized}
}}


