% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ECSFargateProvider.R
\name{ECSFargateProvider}
\alias{ECSFargateProvider}
\title{Create an ECS Fargate cloud provider}
\usage{
ECSFargateProvider(
  clusterName = "R-worker-cluster",
  serverTaskDefName = "R-server-task-definition",
  workerTaskDefName = "R-worker-task-definition",
  securityGroupName = "R-parallel-security-group",
  vpcId = NULL,
  subnetId = NULL,
  securityGroupId = NULL,
  internetGatewayId = NULL,
  routeTableId = NULL,
  workerPublicIpEnable = TRUE
)
}
\arguments{
\item{clusterName}{Character, the cluster name in ECS Fargate}

\item{serverTaskDefName, workerTaskDefName}{Character, the task defintion name for
the server and worker}

\item{securityGroupName}{Character, the security group name}

\item{vpcId, subnetId, securityGroupId, internetGatewayId, routeTableId}{The ID of
the network settings for the container}

\item{workerPublicIpEnable}{Logical, whether to enable the public IP for the worker.
If this value is \code{FALSE}, the worker will not be able to pull the container image from the
public network}
}
\description{
Create an ECS Fargate cloud provider. Note that the arguments for the function
are the settings for ECS fargate. You can find them in the AWS ECS console.
}
