% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/method-default.R
\name{getExportedNames}
\alias{getExportedNames}
\alias{getExportedObject}
\alias{getExportedNames,ANY-method}
\alias{getExportedObject,ANY-method}
\title{Get the exported method and variable from the provider or container}
\usage{
getExportedNames(x)

getExportedObject(x, name)

\S4method{getExportedNames}{ANY}(x)

\S4method{getExportedObject}{ANY}(x, name)
}
\arguments{
\item{x}{A cloud provider or container object}
}
\value{
getExportedNames: The names of the exported methods or variables
getExportedObject: The exported method or variable
}
\description{
Get the exported method and variable from the provider or container. These
methods should be used by the developer to export their APIs to the user. The
\code{DockerCluster} object will call \code{getExportedNames} and \code{getExportedObject} to
export them to the user.
}
\details{
If the exported object is a function, the function can find the \code{DockerCluster} cluster
which contains the cloud provider or the container by the variable \code{cluster}. This
can be useful if the developer needs to change anything in the cluster besides the provider
or container itself. If the argument \code{cluster} is provided as a function argument,
it will be removed when it is exported to the user. Therefore, you can explicitly
define the argument \code{cluster} and use it in your function. Users would not be bothered
with the redundant \code{cluster} argument.
}
