% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{generalDockerClusterTest}
\alias{generalDockerClusterTest}
\title{The general testthat function for testing the cluster}
\usage{
generalDockerClusterTest(
  cloudProvider,
  workerContainer,
  workerNumber = 3L,
  testReconnect = TRUE
)
}
\arguments{
\item{cloudProvider}{The CloudProvider}

\item{workerContainer}{The workerContainer}

\item{workerNumber}{The number of workers used in the unit test}

\item{testReconnect}{Logical, whether to test the reconnect feature}
}
\value{
No return value
}
\description{
The general testthat function for testing the cluster. The function is
generally called by the cloud provider to test the functions in the provider.
if \code{testReconnect} is \code{TRUE},
The provider must define \code{reconnectDockerCluster} for making the test function work.
}
