% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{generics-commonParams}
\alias{generics-commonParams}
\title{commom params}
\arguments{
\item{verbose}{Integer. The verbose level, default 1.}

\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{container}{S4 \code{DockerContainer} Object.}

\item{hardware}{S4 \code{DockerHardware} Object.}
}
\value{
No return value
}
\description{
commom params
}
