% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ECSFargateProvider.R
\name{getDockerInstanceIps,ECSFargateProvider-method}
\alias{getDockerInstanceIps,ECSFargateProvider-method}
\title{Get the instance public/private IPs}
\usage{
\S4method{getDockerInstanceIps}{ECSFargateProvider}(provider, instanceHandles, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{instanceHandles}{List. A list of instance handles.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A data.frame with publicIp and privateIp columns and each row corresponds to
an element in instanceHandles.
}
\description{
Get the instance public/private IPs
}
