% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ECSFargateProvider.R
\name{getDockerInstanceStatus,ECSFargateProvider-method}
\alias{getDockerInstanceStatus,ECSFargateProvider-method}
\title{Get the instance status}
\usage{
\S4method{getDockerInstanceStatus}{ECSFargateProvider}(provider, instanceHandles, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{instanceHandles}{List. A list of instance handles.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A character vector with each element corresponding to an instance in instanceHandles.
Each element must be one of three possible characters "initializing", "running" or "stopped"
}
\description{
Get the instance status
}
