% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyProvider.R
\name{getDockerServerIp,DummyProvider-method}
\alias{getDockerServerIp,DummyProvider-method}
\title{Create a Dummy provider for testing the container}
\usage{
\S4method{getDockerServerIp}{DummyProvider}(provider, cluster, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
This function always returns
\code{list(publicIp = "8.8.8.8", publicPort = 123, privateIp = "192.168.1.1", privatePort = 456)}
}
