% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics-CloudProvider.R,
%   R/method-default.R
\name{getDockerWorkerNumber}
\alias{getDockerWorkerNumber}
\alias{getDockerWorkerNumber,ANY-method}
\title{Get the worker number on the cloud}
\usage{
getDockerWorkerNumber(provider, cluster, verbose)

\S4method{getDockerWorkerNumber}{ANY}(provider, cluster, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}

\item{status}{Character(1), the status of the container.}
}
\value{
An integer vector with two elements.
}
\description{
Get the worker number on the cloud. Return an integer vector with two elements,
which are the number of initializing and running workers. The default method will
return the vector \code{c(0L, .getExpectedWorkerNumber(cluster))}
}
\section{Functions}{
\itemize{
\item \code{getDockerWorkerNumber,ANY-method}: The default getDockerWorkerNumber method. Return \code{c(0L, .getExpectedWorkerNumber(cluster))}
}}

