% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyProvider.R
\name{getDockerWorkerNumbers,DummyProvider-method}
\alias{getDockerWorkerNumbers,DummyProvider-method}
\title{Create a Dummy provider for testing the container}
\usage{
\S4method{getDockerWorkerNumbers}{DummyProvider}(provider, cluster, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
This function returns value defined by the environment variable \code{dummyProviderWorkerNumber}
}
