% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{getInstanceIps}
\alias{getInstanceIps}
\title{Get the instance public/private IPs}
\usage{
getInstanceIps(provider, instanceHandles, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{instanceHandles}{List. A list of instance handles.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A data.frame with \code{publicIp} and \code{privateIp} columns
}
\description{
Get the instance public/private IPs. The IPs will be used by the cluster to
make connections between server and worker, server and client. If the instance
does not have the public or private IP, the return value can be the character
"".
}
