% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BiocFERContainer.R
\name{getServerContainer,BiocFERContainer-method}
\alias{getServerContainer,BiocFERContainer-method}
\title{Get the Bioconductor foreach Redis container}
\usage{
\S4method{getServerContainer}{BiocFERContainer}(workerContainer)
}
\value{
A \code{BiocFERContainer} server container
}
\description{
Get the Bioconductor foreach Redis container from the worker container
}
