% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/method-BiocFERContainer.R,
%   R/method-default.R
\name{getServerContainer}
\alias{getServerContainer}
\alias{getServerContainer,BiocFERContainer-method}
\alias{getServerContainer,ANY-method}
\title{Get the server container from the worker container}
\usage{
getServerContainer(container, ...)

\S4method{getServerContainer}{BiocFERContainer}(container, ...)

\S4method{getServerContainer}{ANY}(container, ...)
}
\arguments{
\item{container}{The worker container.}
}
\description{
Get the server container from the worker container
}
\section{Functions}{
\itemize{
\item \code{getServerContainer,BiocFERContainer-method}: Get the Bioconductor foreach Redis container

\item \code{getServerContainer,ANY-method}: The default method, return \code{NULL} value
}}

