% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyProvider.R
\name{getServerStatus,DummyProvider-method}
\alias{getServerStatus,DummyProvider-method}
\title{Create a Dummy provider for testing the container}
\usage{
\S4method{getServerStatus}{DummyProvider}(provider, cluster, verbose)
}
\value{
No return value
}
\description{
This function will return either "running" or "stopped" depending on
the slot \code{isServerRunning}
}
