% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh-key.R
\name{getSSHPubKeyPath}
\alias{getSSHPubKeyPath}
\title{Get the path to the public ssh key}
\usage{
getSSHPubKeyPath()
}
\value{
The path to the public ssh key
}
\description{
Get the path to the public ssh key
}
\examples{
getSSHPubKeyPath()
}
