% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh-key.R
\name{getSSHPubKeyValue}
\alias{getSSHPubKeyValue}
\title{Get the public ssh key}
\usage{
getSSHPubKeyValue()
}
\value{
The public ssh key
}
\description{
Get the public ssh key
}
\examples{
getSSHPubKeyValue()
}
