% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ECSFargateProvider.R
\name{initializeProvider,ECSFargateProvider-method}
\alias{initializeProvider,ECSFargateProvider-method}
\title{Initialize the Fargate provider}
\usage{
\S4method{initializeProvider}{ECSFargateProvider}(provider, cluster, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
Initialize the Fargate provider
}
