% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{killDockerInstances}
\alias{killDockerInstances}
\title{Kill the instances}
\usage{
killDockerInstances(provider, instanceHandles, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{instanceHandles}{List. A list of instance handles.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A logical vector indicating whether the killing operation is success for each instance
}
\description{
Kill the instances. Multiple workers may share the same instance, in this case, all
workers in the same instance should be killed.
}
