% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{ManagedCloudProvider-class}
\alias{ManagedCloudProvider-class}
\alias{.ManagedCloudProvider}
\title{A}
\description{
The root class of the cloud provider
}
\section{Fields}{

\describe{
\item{\code{serverHandle}}{Character(1), the server handle that can be recognized by the
cloud provider.}

\item{\code{workerHandles}}{Character(n), a list object for internal use only.
Please call \code{.getWorkerHandles(cluster)} to access the worker handles}

\item{\code{workerPerHandle}}{An internal counter. Please call \code{.getWorkerHandles(cluster)}
to access the worker handles}
}}


