% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ClusterMethodGetter.R
\name{names,ClusterMethodGetter-method}
\alias{names,ClusterMethodGetter-method}
\title{Get the exported object names}
\usage{
\S4method{names}{ClusterMethodGetter}(x)
}
\arguments{
\item{x}{\code{ClusterMethodGetter} object}
}
\value{
A vector of object names
}
\description{
Get the exported object names
}
