% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-DockerCluster.R
\name{names,DockerCluster-method}
\alias{names,DockerCluster-method}
\title{Show the exported object names}
\usage{
\S4method{names}{DockerCluster}(x)
}
\arguments{
\item{x}{The \code{DockerCluster} object}
}
\value{
A character vector
}
\description{
Show the exported object names
}
