% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ECSFargateProvider.R
\name{reconnectDockerCluster,ECSFargateProvider-method}
\alias{reconnectDockerCluster,ECSFargateProvider-method}
\title{Reconnect the cluster}
\usage{
\S4method{reconnectDockerCluster}{ECSFargateProvider}(provider, cluster, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
Reconnect the cluster. This function will check the instances in
the ECS cluster defined by \code{provider$clusterName} and find if there is any
instance that has the same job queue name as \code{.getJobQueueName(cluster)}. If
so, the function can reconnect the cluster.
}
