% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BiocFERContainer.R
\name{registerParallelBackend,BiocFERContainer-method}
\alias{registerParallelBackend,BiocFERContainer-method}
\title{Register the foreach parallel backend}
\usage{
\S4method{registerParallelBackend}{BiocFERContainer}(container, cluster, verbose = FALSE, ...)
}
\arguments{
\item{container}{The worker container.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
Register the foreach parallel backend
}
