% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyProvider.R
\name{resetDummyProvider}
\alias{resetDummyProvider}
\title{reset the dummy provider}
\usage{
resetDummyProvider()
}
\value{
No return value
}
\description{
reset the dummy provider and remove all the environment variables it
defined.
}
\examples{
resetDummyProvider()
}
