% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyProvider.R
\name{runDockerServer,DummyProvider-method}
\alias{runDockerServer,DummyProvider-method}
\title{Create a Dummy provider for testing the container}
\usage{
\S4method{runDockerServer}{DummyProvider}(provider, cluster, container, hardware, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{container}{S4 \code{DockerContainer} Object. The server container.}

\item{hardware}{S4 \code{DockerHardware} Object. The server hardware.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
This function will set the slot \code{isServerRunning} to \code{TRUE} and \code{cleanup} to \code{FALSE}.
It also adds the environment variable \code{dummyProvider} and \code{dummyProviderClusterData}.
}
