% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ECSFargateProvider.R
\name{runDockerServer,ECSFargateProvider-method}
\alias{runDockerServer,ECSFargateProvider-method}
\title{Run the server container}
\usage{
\S4method{runDockerServer}{ECSFargateProvider}(provider, cluster, container, hardware, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{container}{S4 \code{DockerContainer} Object. The server container.}

\item{hardware}{S4 \code{CloudHardware} Object. The server hardware.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
The server handle in character
}
\description{
Run the server and return the server instance handle.
The function will set the environment variable \code{ECSFargateCloudConfigInfo} to the container
}
