% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ECSFargateProvider.R
\name{runDockerWorkers,ECSFargateProvider-method}
\alias{runDockerWorkers,ECSFargateProvider-method}
\title{Run the worker container}
\usage{
\S4method{runDockerWorkers}{ECSFargateProvider}(
  provider,
  cluster,
  container,
  hardware,
  workerNumber,
  verbose = 0L
)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{container}{S4 \code{DockerContainer} Object. The worker container.}

\item{hardware}{S4 \code{CloudHardware} Object. The worker hardware.}

\item{workerNumber}{Integer. The number of workers needs to be run.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A list of worker handle in character
}
\description{
Run the workers and return a list of worker instance handles.
The function will set the environment variable \code{ECSFargateCloudJobQueueName},
\code{ECSFargateCloudServerIP} and \code{ECSFargateCloudWorkerNumber} to the container
}
