% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics-ManagedCloudProvider.R
\name{runDockerWorkers}
\alias{runDockerWorkers}
\title{Run the worker container}
\usage{
runDockerWorkers(provider, cluster, container, hardware, workerNumber, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{container}{S4 \code{DockerContainer} Object. The worker container.}

\item{hardware}{S4 \code{DockerHardware} Object. The worker hardware.}

\item{workerNumber}{Integer. The number of workers needs to be run.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A character vector with each element corresponding to a worker container.
The length must be equal to \code{workerNumber}
}
\description{
Run the workers and return a character vector of the container handles
which the workers live in.
The handle must be recognized by the cloud provider to manage the containers.
The instance handles can be duplicated if multiple workers share the same container instance.
There is no default method for this generic.
}
