% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{runServer}
\alias{runServer}
\title{Run the server}
\usage{
runServer(provider, cluster, container, hardware, verbose)
}
\arguments{
\item{provider}{S4 CloudProvider object. The service provider.}

\item{cluster}{S4 DockerCluster object.}

\item{container}{S4 Container Object. The server container.}

\item{hardware}{S4 CloudHardware Object. The server hardware.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
Any object that can be used by the cluster to identify the server instance.
}
\description{
Run the server and return the server's instance handle.
}
\section{Instance Handle}{

Any data type can be used as the instance handle as long as it supports
\code{identical} and \code{unique} functions. The \code{DockerParallel} object will use the
handle to check the instance status or kill the instance. It is recommended
to use \code{numeric} or \code{character} as the instance handle.
}

