% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{runWorkers}
\alias{runWorkers}
\title{Run the workers}
\usage{
runWorkers(provider, cluster, container, hardware, workerNumber, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{container}{S4 Container Object. The worker container.}

\item{hardware}{S4 CloudHardware Object. The worker hardware.}

\item{workerNumber}{Integer. The number of workers need to be run.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A list of object that can be used by the cluster to identify the worker instances.
}
\description{
Run the workers and return a list of worker's instance handles.
}
\section{Instance Handle}{

Any data type can be used as the instance handle as long as it supports
\code{identical} and \code{unique} functions. The \code{DockerParallel} object will use the
handle to check the instance status or kill the instance. It is recommended
to use \code{numeric} or \code{character} as the instance handle.
}

