% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-export.R
\name{serializeDockerClusterStaticData}
\alias{serializeDockerClusterStaticData}
\alias{unserializeDockerClusterStaticData}
\title{Serialize docker cluster static data}
\usage{
serializeDockerClusterStaticData(cluster)

unserializeDockerClusterStaticData(cluster, serializedData)
}
\arguments{
\item{cluster}{The \code{DockerCluster} object}

\item{serializedData}{The serialized data returned by \code{serializeDockerClusterStaticData}}
}
\value{
serializeDockerClusterStaticData: The binary serialized data
unserializeDockerClusterStaticData: No return value should be expected, the cluster
object will be updated.
}
\description{
Serialize docker cluster static data. This function is designed for the
generic \code{reconnectDockerCluster}. The serialized data can be used by the cloud
provider to recover the \code{DockerCluster} object.
}
