% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyProvider.R
\name{setDockerWorkerNumber,DummyProvider-method}
\alias{setDockerWorkerNumber,DummyProvider-method}
\title{Create a Dummy provider for testing the container}
\usage{
\S4method{setDockerWorkerNumber}{DummyProvider}(
  provider,
  cluster,
  container,
  hardware,
  workerNumber,
  verbose
)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{container}{S4 \code{DockerContainer} Object.}

\item{hardware}{S4 \code{DockerHardware} Object.}

\item{workerNumber}{Integer(1), the number of the workers.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
This function will set the environment variable \code{dummyProviderWorkerNumber} and stores its
container in the slot \code{workerContainer}.
}
