% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics-CloudProvider.R,
%   R/method-ManagedCloudProvider.R
\name{setDockerWorkerNumber}
\alias{setDockerWorkerNumber}
\alias{setDockerWorkerNumber,ManagedCloudProvider-method}
\title{Set the worker number on the cloud. There is no default method for this generic.}
\usage{
setDockerWorkerNumber(
  provider,
  cluster,
  container,
  hardware,
  workerNumber,
  verbose
)

\S4method{setDockerWorkerNumber}{ManagedCloudProvider}(
  provider,
  cluster,
  container,
  hardware,
  workerNumber,
  verbose
)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{container}{S4 \code{DockerContainer} Object.}

\item{hardware}{S4 \code{DockerHardware} Object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
The worker number on the cloud
}
\description{
Set the worker number on the cloud. The provider needs to scale the worker
number up and down accordingly.
}
\section{Functions}{
\itemize{
\item \code{setDockerWorkerNumber,ManagedCloudProvider-method}: The method for the \code{ManagedCloudProvider}
}}

