% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DockerParallel-ssh-key.R
\name{setSSHPubKey}
\alias{setSSHPubKey}
\title{Set the ssh key file}
\usage{
setSSHPubKey(publicKey = NULL)
}
\arguments{
\item{publicKey}{path to the public key}
}
\description{
Set the ssh key file. This function will be called when the package is
loaded. If no argument is provided, it
will look at the environment variables \code{DockerParallelSSHPublicKey}
}
