% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-ClusterMethodGetter.R
\name{show,ClusterMethodGetter-method}
\alias{show,ClusterMethodGetter-method}
\title{print method}
\usage{
\S4method{show}{ClusterMethodGetter}(object)
}
\arguments{
\item{object}{\code{ClusterMethodGetter} object}
}
\value{
No return value
}
\description{
print method
}
