% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DockerHardware.R
\name{show,DockerHardware-method}
\alias{show,DockerHardware-method}
\title{Print the docker hardware}
\usage{
\S4method{show}{DockerHardware}(object)
}
\arguments{
\item{object}{The DockerHardware object}
}
\value{
No return value
}
\description{
Print the docker hardware
}
\examples{
## This is the internal function and should not be called by the user
## It is for illustration purpose only
hardware <- DockerParallel:::DockerHardware()
show(hardware)
}
