% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{validateContainer}
\alias{validateContainer}
\title{Validate the container}
\usage{
validateContainer(provider, container)
}
\arguments{
\item{provider}{S4 CloudProvider object. The service provider.}

\item{container}{S4 Container Object.}
}
\value{
A logical value
}
\description{
Check if the current container provider is supported by the service provider
}
