% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-export.R
\name{waitInstanceUntilRunning}
\alias{waitInstanceUntilRunning}
\title{Wait the instances until they are running}
\usage{
waitInstanceUntilRunning(
  provider,
  instanceHandles,
  progressBar = FALSE,
  waitTime = 1
)
}
\value{
a logical value indicating if the instances are running
}
\description{
Wait the instances until they are running
}
