% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics-ManagedCloudProvider.R,
%   R/method-default-ManagedCloudProvider.R
\name{getDockerWorkerStatus}
\alias{getDockerWorkerStatus}
\alias{IsDockerWorkerInitializing}
\alias{IsDockerWorkerRunning}
\alias{IsDockerWorkerStopped}
\alias{getDockerWorkerStatus,ANY-method}
\alias{IsDockerWorkerInitializing,ANY-method}
\alias{IsDockerWorkerRunning,ANY-method}
\alias{IsDockerWorkerStopped,ANY-method}
\title{Get the worker status}
\usage{
getDockerWorkerStatus(provider, workerHandles, verbose)

IsDockerWorkerInitializing(provider, workerHandles, verbose)

IsDockerWorkerRunning(provider, workerHandles, verbose)

IsDockerWorkerStopped(provider, workerHandles, verbose)

\S4method{getDockerWorkerStatus}{ANY}(provider, workerHandles, verbose = 0L)

\S4method{IsDockerWorkerInitializing}{ANY}(provider, workerHandles, verbose = 0L)

\S4method{IsDockerWorkerRunning}{ANY}(provider, workerHandles, verbose = 0L)

\S4method{IsDockerWorkerStopped}{ANY}(provider, workerHandles, verbose = 0L)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{workerHandles}{Character(n). A character vector of instance handles.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
\code{getDockerWorkerStatus} : A character vector with each element corresponding
to an instance in \code{workerHandles}. Each element must be one of three possible characters
\code{"initializing"}, \code{"running"} or \code{"stopped"}

\code{IsDockerWorkerInitializing}, \code{IsDockerWorkerRunning}, \code{IsDockerWorkerStopped}:
A logical vector with each element corresponding to the status of each instance
}
\description{
Get the worker status. Unless you have a faster implementation, you only need to
define \code{getDockerWorkerStatus}. The function should return a character vector with
each element corresponding to a worker in \code{workerHandles}.
Each element must be one of three possible characters \code{"initializing"}, \code{"running"} or
\code{"stopped"}. The default \code{getDockerWorkerStatus} returns
a vector of character \code{"running"} with the same length of the input worker handles.
}
