% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{.getCloudProvider}
\alias{.getCloudProvider}
\alias{.getCloudConfig}
\alias{.getServerContainer}
\alias{.getWorkerContainer}
\alias{.getCloudRuntime}
\alias{.getClusterSettings}
\alias{.getVerbose}
\alias{.getStopClusterOnExit}
\alias{.setCloudProvider}
\alias{.setCloudConfig}
\alias{.setServerContainer}
\alias{.setWorkerContainer}
\alias{.setCloudRuntime}
\alias{.setClusterSettings}
\alias{.setVerbose}
\alias{.setStopClusterOnExit}
\alias{.getJobQueueName}
\alias{.getExpectedWorkerNumber}
\alias{.getWorkerHardware}
\alias{.getServerHardware}
\alias{.getServerWorkerSameLAN}
\alias{.getServerClientSameLAN}
\alias{.getServerPassword}
\alias{.getServerPort}
\alias{.setJobQueueName}
\alias{.setExpectedWorkerNumber}
\alias{.setWorkerHardware}
\alias{.setServerHardware}
\alias{.setServerWorkerSameLAN}
\alias{.setServerClientSameLAN}
\alias{.setServerPassword}
\alias{.setServerPort}
\alias{.getServerFromOtherSource}
\alias{.getServerPrivateIp}
\alias{.getServerPrivatePort}
\alias{.getServerPublicIp}
\alias{.getServerPublicPort}
\alias{.getInitializingWorkerNumber}
\alias{.getRunningWorkerNumber}
\alias{.setServerPrivateIp}
\alias{.setServerPublicIp}
\alias{.setServerPrivatePort}
\alias{.setServerPublicPort}
\alias{.setInitializingWorkerNumber}
\alias{.setRunningWorkerNumber}
\alias{.setServerFromOtherSource}
\title{Accessor functions}
\usage{
.getCloudProvider(cluster)

.getCloudConfig(cluster)

.getServerContainer(cluster)

.getWorkerContainer(cluster)

.getCloudRuntime(cluster)

.getClusterSettings(cluster)

.getVerbose(cluster)

.getStopClusterOnExit(cluster)

.setCloudProvider(cluster, value)

.setCloudConfig(cluster, value)

.setServerContainer(cluster, value)

.setWorkerContainer(cluster, value)

.setCloudRuntime(cluster, value)

.setClusterSettings(cluster, value)

.setVerbose(cluster, value)

.setStopClusterOnExit(cluster, value)

.getJobQueueName(cluster)

.getExpectedWorkerNumber(cluster)

.getWorkerHardware(cluster)

.getServerHardware(cluster)

.getServerWorkerSameLAN(cluster)

.getServerClientSameLAN(cluster)

.getServerPassword(cluster)

.getServerPort(cluster)

.setJobQueueName(cluster, value)

.setExpectedWorkerNumber(cluster, value)

.setWorkerHardware(cluster, value)

.setServerHardware(cluster, value)

.setServerWorkerSameLAN(cluster, value)

.setServerClientSameLAN(cluster, value)

.setServerPassword(cluster, value)

.setServerPort(cluster, value)

.getServerFromOtherSource(cluster)

.getServerPrivateIp(cluster)

.getServerPrivatePort(cluster)

.getServerPublicIp(cluster)

.getServerPublicPort(cluster)

.getInitializingWorkerNumber(cluster)

.getRunningWorkerNumber(cluster)

.setServerPrivateIp(cluster, value)

.setServerPublicIp(cluster, value)

.setServerPrivatePort(cluster, value)

.setServerPublicPort(cluster, value)

.setInitializingWorkerNumber(cluster, value)

.setRunningWorkerNumber(cluster, value)

.setServerFromOtherSource(cluster, value)
}
\arguments{
\item{cluster}{A \code{DockerCluster} object}

\item{value}{The value you want to set/add/remove}
}
\value{
No return value for the setter. The getter will get the object from the cluster.
}
\description{
Accessor functions for the developer.
}
