% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{CloudConfig-class}
\alias{CloudConfig-class}
\alias{.CloudConfig}
\title{The cloud configuration}
\description{
The cloud configuration
}
\section{Slots}{

\describe{
\item{\code{jobQueueName}}{The name of the job queue}

\item{\code{workerNumber}}{The required number of workers that should be
run on the cloud}

\item{\code{serverHardware}}{The server hardware}

\item{\code{workerHardware}}{The worker hardware}
}}


