% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{CloudConfig-class}
\alias{CloudConfig-class}
\alias{.CloudConfig}
\title{The cloud configuration}
\description{
The cloud configuration. It is a class purely for storing the information
for the cloud.
The values in \code{CloudConfig} in a cluster can be accessed by the getter function
which starts with the prefix \code{.get}(e.g. \code{.getJobQueueName(cluster)}).
}
\section{Fields}{

\describe{
\item{\code{jobQueueName}}{Character(1), the name of the job queue.}

\item{\code{expectedWorkerNumber}}{Integer(1), the expected number of workers that should be
run on the cloud.}

\item{\code{serverHardware}}{DockerHardware, the server hardware.}

\item{\code{workerHardware}}{DockerHardware, the worker hardware.}

\item{\code{serverPort}}{Integer(1) or integer(0), the port that will be used by the worker
to connect with the server.}

\item{\code{serverPassword}}{Character(1) or character(0), the server password.}

\item{\code{serverWorkerSameLAN}}{Logical(1), whether the server and workers are behind
the same router.}

\item{\code{serverClientSameLAN}}{Logical(1), whether the server and client are behind
the same router.}
}}


