% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-export.R
\name{clusterPreset}
\alias{clusterPreset}
\title{Set the default cloud provider and container}
\usage{
clusterPreset(
  cloudProvider = c("", "ECSFargateProvider"),
  container = c("", "rbaseDoRedis", "rbaseRedisParam", "biocDoRedis", "biocRedisParam")
)
}
\arguments{
\item{cloudProvider}{The default cloud provider name, can be abbreviated}

\item{container}{The default container name, can be abbreviated}
}
\value{
No return value
}
\description{
Set the default cloud provider and container. You must install the provider and container
packages before using them.
}
\examples{
\dontrun{
clusterPreset(cloudProvider = "ECSFargateProvider", container = "rbaseDoRedis")
cluster <- makeDockerCluster()
cluster
}
}
