% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/method-BiocFERContainer.R
\name{configServerContainerEnv}
\alias{configServerContainerEnv}
\alias{configServerContainerEnv,BiocFERContainer-method}
\title{Configurate the server container environment}
\usage{
configServerContainerEnv(container, cluster, verbose)

\S4method{configServerContainerEnv}{BiocFERContainer}(container, cluster, verbose = FALSE)
}
\arguments{
\item{container}{Reference Container Object. The server container.}

\item{cluster}{S4 DockerCluster object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A \code{container} object
}
\description{
Configurate the server container environment. Developers can use this function
to set the server password, port number and etc. via the container environment variable.
}
\section{Functions}{
\itemize{
\item \code{configServerContainerEnv,BiocFERContainer-method}: configure the server container
environment, only the environment \code{serverPort}, \code{serverPassword} and \code{sshPubKey}
will be set in this method
}}

