% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/method-BiocFERContainer.R
\name{configWorkerContainerEnv}
\alias{configWorkerContainerEnv}
\alias{configWorkerContainerEnv,BiocFERContainer-method}
\title{Configurate the worker container environment}
\usage{
configWorkerContainerEnv(container, cluster, workerNumber, verbose)

\S4method{configWorkerContainerEnv}{BiocFERContainer}(container, cluster, workerNumber, verbose = FALSE)
}
\arguments{
\item{container}{Reference Container Object. The worker container.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{workerNumber}{Integer. The number of workers in a container.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
An object which has the same class as \code{container}
}
\description{
Configurate the worker container environment. Developers can use this function
to set the server Ip, password and etc. via the container environment variable.
}
\section{Functions}{
\itemize{
\item \code{configWorkerContainerEnv,BiocFERContainer-method}: configure the worker container
}}

