% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics-CloudProvider.R,
%   R/method-default.R
\name{dockerClusterExists}
\alias{dockerClusterExists}
\alias{dockerClusterExists,ANY-method}
\title{Whether the cluster is running on the cloud?}
\usage{
dockerClusterExists(provider, cluster, verbose)

\S4method{dockerClusterExists}{ANY}(provider, cluster, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A logical value
}
\description{
The function checks whether the cluster is running on the cloud. It returns
\code{TRUE} if the cluster specific to the value from \code{.getJobQueueName(cluster)} exists.
The default method always returns \code{FALSE}
}
\section{Functions}{
\itemize{
\item \code{dockerClusterExists,ANY-method}: The default method, it always returns \code{FALSE}.
}}

