% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{DockerContainer-class}
\alias{DockerContainer-class}
\alias{.DockerContainer}
\title{The root class of the container}
\description{
The root class of the container
}
\section{Fields}{

\describe{
\item{\code{name}}{Character(1) or character(0), the optional name of a container.}

\item{\code{backend}}{Character(1), the backend used by the parallel package}

\item{\code{maxWorkerNum}}{Integer(1), the maximum worker number in a container.}

\item{\code{environment}}{List, the environment variables in the container.}

\item{\code{image}}{Character(1), the container image.}
}}


