% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{DockerHardware-class}
\alias{DockerHardware-class}
\alias{.DockerHardware}
\title{The hardware for running the docker}
\description{
The hardware for running the docker
}
\section{Slots}{

\describe{
\item{\code{cpu}}{Numeric(1), the CPU limitation for the docker. 1024 CPU unit
corresponds to 1 core.}

\item{\code{memory}}{Numeric(1), the memory limitation for the docker, the unit
is MB.}

\item{\code{id}}{character(1) or character(0), the id of the hardware, the meaning of
\code{id} depends on the cloud provider.}
}}

