% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyProvider.R
\name{DummyProvider}
\alias{DummyProvider}
\title{Create a Dummy provider for testing the container}
\usage{
DummyProvider(initialized = FALSE, isServerRunning = FALSE, cleanup = FALSE)
}
\arguments{
\item{initialized, isServerRunning, cleanup}{logical(1), the flags}
}
\value{
A \code{DummyProvider} object
}
\description{
Create a Dummy provider for testing the container
}
\examples{
DummyProvider()
}
