% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecs_credential.R
\name{ecs_set_credentials}
\alias{ecs_set_credentials}
\title{Setting up AWS credentials}
\usage{
ecs_set_credentials(
  key_file = NULL,
  access_key_id = NULL,
  secret_access_key = NULL,
  region = NULL,
  profile = NULL
)
}
\arguments{
\item{key_file}{The csv credential file that is downloaded from AWS}
}
\description{
Setting up AWS credentials. For details, see \code{?aws.signature::locate_credentials}
}
