% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R
\name{getInstanceIps}
\alias{getInstanceIps}
\title{Get the cluster IP}
\usage{
getInstanceIps(provider, instanceHandles, verbose)
}
\arguments{
\item{serverHandle}{the server handles}
}
\value{
A data.frame with \code{publicIp} and \code{privateIp} columns
}
\description{
Get the cluster IP, the IP can be used for the worker to connect with
the server. The IP can be anything and does not have to have any meaning
to the cluster.
}
