% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/method-default.R
\name{getServerContainer}
\alias{getServerContainer}
\alias{getServerContainer,ANY-method}
\title{Get the server container from the worker container}
\usage{
getServerContainer(workerContainer)

\S4method{getServerContainer}{ANY}(workerContainer)
}
\arguments{
\item{workerContainer}{The worker container.}
}
\value{
A server container
}
\description{
Get the server container from the worker container. This function will
be called by the \code{DockerCluster} object when the user only provides a worker container
to its constructor.
There is no default method defined for this generic.
}
\section{Functions}{
\itemize{
\item \code{getServerContainer,ANY-method}: The default method, return \code{NULL} value
}}

