% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics-CloudProvider.R
\name{getServerStatus}
\alias{getServerStatus}
\title{Get the server status}
\usage{
getServerStatus(provider, cluster, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
Character(1)
}
\description{
Get the server status, return a character value which must be in one of three values
\code{"initializing"}, \code{"running"} or \code{"stopped"}. The default method always returns \code{"running"}
}
