% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/method-default.R
\name{getInstanceStatus}
\alias{getInstanceStatus}
\alias{IsInstanceInitializing}
\alias{IsInstanceRunning}
\alias{IsInstanceStopped}
\alias{getInstanceStatus,ANY-method}
\alias{IsInstanceInitializing,ANY-method}
\alias{IsInstanceRunning,ANY-method}
\alias{IsInstanceStopped,ANY-method}
\title{Get the instance status}
\usage{
getInstanceStatus(provider, instanceHandles, verbose)

IsInstanceInitializing(provider, instanceHandles, verbose)

IsInstanceRunning(provider, instanceHandles, verbose)

IsInstanceStopped(provider, instanceHandles, verbose)

\S4method{getInstanceStatus}{ANY}(provider, instanceHandles, verbose = FALSE)

\S4method{IsInstanceInitializing}{ANY}(provider, instanceHandles, verbose = FALSE)

\S4method{IsInstanceRunning}{ANY}(provider, instanceHandles, verbose = FALSE)

\S4method{IsInstanceStopped}{ANY}(provider, instanceHandles, verbose = FALSE)
}
\arguments{
\item{provider}{S4 CloudProvider object. The service provider.}

\item{instanceHandles}{List. A list of instance handles.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
\code{getInstanceStatus} : A character vector with each element corresponding
to an instance in \code{instanceHandles}. Each element must be one of three possible characters
\code{"initializing"}, \code{"running"} or \code{"stopped"}

\code{IsInstanceInitializing}, \code{IsInstanceRunning}, \code{IsInstanceStopped}:
A logical vector with each element corresponding to the status of each instance
}
\description{
Get the instance status. Unless you have a special requirement, you only need to
define \code{getInstanceStatus}. The default \code{getInstanceStatus} do nothing but return
a vector of "running" with the same lenght of the input instance handles.
}
