% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics-CloudProvider.R
\name{setDockerWorkerNumber}
\alias{setDockerWorkerNumber}
\title{Set the worker number on the cloud. There is no default method for this generic.}
\usage{
setDockerWorkerNumber(
  provider,
  cluster,
  container,
  hardware,
  workerNumber,
  verbose
)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{container}{S4 \code{DockerContainer} Object.}

\item{hardware}{S4 \code{DockerHardware} Object.}

\item{workerNumber}{Integer(1), the number of the workers.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
Set the worker number on the cloud. The provider needs to scale the worker
number up and down accordingly.
}
