% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-export.R
\name{waitInstanceUntilRunning}
\alias{waitInstanceUntilRunning}
\title{Wait the instances until they have the running status}
\usage{
waitInstanceUntilRunning(
  provider,
  instanceHandles,
  progressBar = FALSE,
  checkInterval = 1,
  maxWaitTime = 60 * 5
)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{instanceHandles}{List. A list of instance handles.}

\item{progressBar}{Logical, whether to show a progress bar while waiting}

\item{checkInterval}{Numeric, the time interval in seconds between two status checks.}

\item{maxWaitTime}{Numeric, the maximum wait time in seconds. There will be no error
if the wait time exceeds the maximum wait time.}
}
\value{
a logical vector indicating if the instances are running
}
\description{
Wait the instances until they have the running status
}
